refactor: 데이터레이어의 repositories를 repository 로 변경
