+++ data/src/main/java/com/paykidscompose/data/repository/QuestRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.paykidscompose.data.repositories
+package com.paykidscompose.data.repository
 
 import com.paykidscompose.common.exception.PayKidsException
 import com.paykidscompose.common.model.quest.QuestModel
